Convert test/plots/* to TSX (unit 8: metro-unemployment-moving…overlapping-area)
Convert 22 React test fixtures from createElement calls to JSX. Drops
React default imports where only used for createElement, preserves named
imports and type usage. Snapshot output unchanged.

Note: ordinal-opacity.ts and overlapping-area.ts were not present in the
working tree, so the unit covers 22 of 24 nominally listed files.
