+++ test/plots/mixed-facets.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {Plot, BarY, RuleY} from "../../src/react/index.js";
 
 export async function mixedFacets() {
@@ -11,9 +10,11 @@ export async function mixedFacets() {
     {date: new Date("2024-02-01"), name: "b", value: 2},
     {date: new Date("2024-02-01"), name: "a", value: 3}
   ];
-  return React.createElement(Plot, {},
-    React.createElement(BarY, {data, x: "name", y: "value", fill: "name", fx: "date", fy: "name"}),
-    React.createElement(BarY, {data, x: "name", y: "value", fx: "date", stroke: "currentColor"}),
-    React.createElement(RuleY, {data: [0]})
+  return (
+    <Plot>
+      <BarY data={data} x="name" y="value" fill="name" fx="date" fy="name" />
+      <BarY data={data} x="name" y="value" fx="date" stroke="currentColor" />
+      <RuleY data={[0]} />
+    </Plot>
   );
 }
