+++ test/plots/ordinal-bar.tsx
@@ -1,0 +1,14 @@
+import {Plot, BarY, RuleY} from "../../src/react/index.js";
+
+export async function ordinalBar() {
+  return (
+    <Plot
+      y={{
+        grid: true
+      }}
+    >
+      <BarY data="ABCDEF" />
+      <RuleY data={[0]} />
+    </Plot>
+  );
+}
